removed logs args, tightened types for session state file
