+++ src/EXPERIMENTAL__infra-as-code/session.ts
@@ -1,28 +1,34 @@
-import { readFile, writeFile } from "node:fs/promises"
+import { mkdir, mkdtemp, readFile, writeFile } from "node:fs/promises"
+import path = require("node:path")
 
 import type { InfraAsCodeApiResult } from "./api"
 import { isEnoentError } from "./utils"
 
-export type InfraAsCodeSessionState = {
+export const DEFAULT_SESSION_STATE_FILE_DIRECTORY = path.resolve(
+  process.cwd(),
+  "tmp",
+  "infra-as-code",
+  "sessions"
+)
+
+export const DEFAULT_SESSION_STATE_FILE_PREFIX = "infra-as-code-session-state"
+
+export type InfraAsCodeSessionInput = {
   existingResources: Record<string, unknown>
   existingProperties: Record<string, string>
 }
 
-type InfraAsCodeWrittenSessionState = {
+type InfraAsCodeSessionOutput = {
   resourceIdToPointerMappings: Record<string, unknown>
   resourceIdToPropertyIdMappings: Record<string, string>
 }
-
-type InfraAsCodeSessionStateFile = Partial<
-  InfraAsCodeSessionState & InfraAsCodeWrittenSessionState
->
 
 /**
  * Reads optional existing resource mappings used to make reruns incremental.
  */
 export async function readSessionState(
   existingResourcesFilePath: string | undefined
-): Promise<InfraAsCodeSessionState> {
+): Promise<InfraAsCodeSessionInput> {
   if (existingResourcesFilePath === undefined) {
     return emptySessionState()
   }
@@ -30,15 +36,11 @@ export async function readSessionState(
   try {
     const parsed = JSON.parse(
       await readFile(existingResourcesFilePath, "utf8")
-    ) as InfraAsCodeSessionStateFile | null
+    )
 
     return {
-      existingResources:
-        parsed?.existingResources ?? parsed?.resourceIdToPointerMappings ?? {},
-      existingProperties:
-        parsed?.existingProperties ??
-        parsed?.resourceIdToPropertyIdMappings ??
-        {},
+      existingResources: parsed?.existingResources ?? {},
+      existingProperties: parsed?.existingProperties ?? {},
     }
   } catch (error) {
     if (isEnoentError(error)) {
@@ -53,10 +55,10 @@ export async function readSessionState(
  */
 export async function writeSessionState(
   sessionStateFilePath: string,
-  priorState: InfraAsCodeSessionState,
+  priorState: InfraAsCodeSessionInput,
   result: InfraAsCodeApiResult
 ): Promise<void> {
-  const nextState: InfraAsCodeWrittenSessionState = {
+  const nextState: InfraAsCodeSessionOutput = {
     resourceIdToPointerMappings: {
       ...priorState.existingResources,
       ...(result.resourceIdToPointerMappings ?? {}),
@@ -75,9 +77,23 @@ export async function writeSessionState(
 }
 
 /**
+ * Creates a unique default path where the next session-state file can be saved.
+ */
+export async function createDefaultSessionStateFilePath(): Promise<string> {
+  await mkdir(DEFAULT_SESSION_STATE_FILE_DIRECTORY, { recursive: true })
+  const tempDir = await mkdtemp(
+    path.join(
+      DEFAULT_SESSION_STATE_FILE_DIRECTORY,
+      `${DEFAULT_SESSION_STATE_FILE_PREFIX}-`
+    )
+  )
+  return path.join(tempDir, "session-state.json")
+}
+
+/**
  * Provides empty mappings when no session-state file exists yet.
  */
-function emptySessionState(): InfraAsCodeSessionState {
+function emptySessionState(): InfraAsCodeSessionInput {
   return {
     existingResources: {},
     existingProperties: {},
